+++ app/lib/articles.server.jsx
@@ -12,27 +12,23 @@ export const getArticles = async () => {
         frontmatter: article.frontmatter,
       };
     });
-    return articles;
-
-    
-
-    // const modules = import.meta.glob<{ frontmatter: Frontmatter }>(
-    //   "../routes/blog.*.mdx",
-    //   { eager: true }
-    // );
-    // const build = await import("virtual:remix/server-build");
-
-
-
-    // const posts = Object.entries(modules).map(([file, post]) => {
-    //   let id = file.replace("../", "").replace(/\.mdx$/, "");
-    //   let slug = build.routes[id].path;
-    //   if (slug === undefined) throw new Error(`No route for ${id}`);
-  
-    //   return {
-    //     slug,
-    //     frontmatter: post.frontmatter,
-    //   };
-    // });
-    // return sortBy(posts, (post) => post.frontmatter.published, "desc");
+    //return articles;
+    return sortBy(articles, (article) => article.frontmatter.published, "desc");
   };
+
+  function sortBy(
+  arr,
+  key,
+  dir
+) {
+  return arr.sort((a, b) => {
+    const res = compare(key(a), key(b));
+    return dir === "asc" ? res : -res;
+  });
+}
+
+function compare(a, b) {
+  if (a < b) return -1;
+  if (a > b) return 1;
+  return 0;
+}