+++ app/routes/articles_._index.jsx
@@ -1,0 +1,41 @@
+//inspired by https://remix-blog-mdx.fly.dev/
+
+import { getArticles } from "../lib/articles.server.jsx";
+import { useLoaderData } from "@remix-run/react";
+import ArticleImage from "../components/article/articleimage.jsx";
+import Navbar from "../components/navbar/navbar.jsx";
+import Footer from "../components/footer.jsx"
+
+export const loader = async () => Response.json(await getArticles());
+
+export default function Index() {
+    const articles = useLoaderData();
+
+    return (
+        <>
+            <Navbar />
+            <div className="">
+                <div className="mx-auto max-w-7xl px-6 lg:px-8">
+                    <div className="mx-auto mt-10 grid max-w-2xl grid-cols-1 gap-x-8 gap-y-16 pt-10 sm:mt-16 sm:pt-16 lg:mx-0 lg:max-w-none lg:grid-cols-2">
+                        {articles.map((article) => (
+                            <article className="flex max-w-xl flex-col items-start  ">
+                                {article.frontmatter.image ? <ArticleImage imageName={article.frontmatter.image} /> : ""}
+                                <div className="group relative">
+                                    <h3 className="mt-3 text-2xl font-semibold text-600 dark:text-500 group-hover:text-gray-400 dark:group-hover:text-gray-600">
+                                        <a href={"/articles/" + article.slug}>
+                                            <span className="absolute inset-0" />
+                                            {article.frontmatter.title}
+                                        </a>
+                                    </h3>
+                                    <p className="mt-5 line-clamp-3 text-lg text-gray-600 dark:text-gray-200">{article.frontmatter.description}</p>
+                                </div>
+                            </article>
+                        ))}
+                    </div>
+                </div>
+            </div>
+            <Footer />
+        </>
+    )
+
+}